Remove database files when logging out or when syncs are removed

+++ src/db/sqlite.cpp
@@ -65,17 +65,24 @@ DbTable* SqliteDbAccess::open(FileSystemAccess* fsaccess, string* name)
         return NULL;
     }
 
-    return new SqliteDbTable(db);
+    return new SqliteDbTable(db, fsaccess, &dbdir);
 }
 
-SqliteDbTable::SqliteDbTable(sqlite3* cdb)
+SqliteDbTable::SqliteDbTable(sqlite3* cdb, FileSystemAccess *fs, string *filepath)
 {
     db = cdb;
     pStmt = NULL;
+    fsaccess = fs;
+    dbfile = *filepath;
 }
 
 SqliteDbTable::~SqliteDbTable()
 {
+    if (!db)
+    {
+        return;
+    }
+
     if (pStmt)
     {
         sqlite3_finalize(pStmt);
@@ -87,6 +94,11 @@ SqliteDbTable::~SqliteDbTable()
 // set cursor to first record
 void SqliteDbTable::rewind()
 {
+    if (!db)
+    {
+        return;
+    }
+
     if (pStmt)
     {
         sqlite3_reset(pStmt);
@@ -100,6 +112,11 @@ void SqliteDbTable::rewind()
 // retrieve next record through cursor
 bool SqliteDbTable::next(uint32_t* index, string* data)
 {
+    if (!db)
+    {
+        return false;
+    }
+
     if (!pStmt)
     {
         return false;
@@ -124,6 +141,11 @@ bool SqliteDbTable::next(uint32_t* index, string* data)
 // retrieve record by index
 bool SqliteDbTable::get(uint32_t index, string* data)
 {
+    if (!db)
+    {
+        return false;
+    }
+
     sqlite3_stmt *stmt;
 
     int rc = sqlite3_prepare(db, "SELECT content FROM statecache WHERE id = ?", -1, &stmt, NULL);
@@ -157,6 +179,11 @@ bool SqliteDbTable::get(uint32_t index, string* data)
 // add/update record by index
 bool SqliteDbTable::put(uint32_t index, char* data, unsigned len)
 {
+    if (!db)
+    {
+        return false;
+    }
+
     sqlite3_stmt *stmt;
 
     int rc = sqlite3_prepare(db, "INSERT OR REPLACE INTO statecache (id, content) VALUES (?, ?)", -1, &stmt, NULL);
@@ -194,6 +221,11 @@ bool SqliteDbTable::put(uint32_t index, char* data, unsigned len)
 // delete record by index
 bool SqliteDbTable::del(uint32_t index)
 {
+    if (!db)
+    {
+        return false;
+    }
+
     char buf[64];
 
     sprintf(buf, "DELETE FROM statecache WHERE id = %" PRIu32, index);
@@ -204,25 +236,66 @@ bool SqliteDbTable::del(uint32_t index)
 // truncate table
 void SqliteDbTable::truncate()
 {
+    if (!db)
+    {
+        return;
+    }
+
     sqlite3_exec(db, "DELETE FROM statecache", 0, 0, NULL);
 }
 
 // begin transaction
 void SqliteDbTable::begin()
 {
+    if (!db)
+    {
+        return;
+    }
+
     sqlite3_exec(db, "BEGIN", 0, 0, NULL);
 }
 
 // commit transaction
 void SqliteDbTable::commit()
 {
+    if (!db)
+    {
+        return;
+    }
+
     sqlite3_exec(db, "COMMIT", 0, 0, NULL);
 }
 
 // abort transaction
 void SqliteDbTable::abort()
 {
+    if (!db)
+    {
+        return;
+    }
+
     sqlite3_exec(db, "ROLLBACK", 0, 0, NULL);
+}
+
+void SqliteDbTable::remove()
+{
+    if (!db)
+    {
+        return;
+    }
+
+    if (pStmt)
+    {
+        sqlite3_finalize(pStmt);
+    }
+    abort();
+    sqlite3_close(db);
+
+    db = NULL;
+
+    string localpath;
+    fsaccess->path2local(&dbfile, &localpath);
+    fsaccess->unlinklocal(&localpath);
 }
 } // namespace
